sdl2: adds cast refs #61
svn merge -r 1885:1886 http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui/WORK_01
svn merge -r 1888:1889 http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui/WORK_01

git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@1893 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ sdl2/ini.c
@@ -208,7 +208,7 @@ const INITBL	*pterm;
 	milstr_ncpy(work, "[", sizeof(work));
 	milstr_ncat(work, title, sizeof(work));
 	milstr_ncat(work, "]\r\n", sizeof(work));
-	file_write(fh, work, strlen(work));
+	file_write(fh, work, (UINT)strlen(work));
 
 	p = tbl;
 	pterm = tbl + count;
@@ -270,9 +270,9 @@ const INITBL	*pterm;
 				break;
 		}
 		if (set == SUCCESS) {
-			file_write(fh, p->item, strlen(p->item));
+			file_write(fh, p->item, (UINT)strlen(p->item));
 			file_write(fh, "=", 1);
-			file_write(fh, work, strlen(work));
+			file_write(fh, work, (UINT)strlen(work));
 			file_write(fh, "\r\n", 2);
 		}
 		p++;
